# docs/api/openapi-v1.yaml
openapi: 3.1.0
info:
  title: Backup Gateway API
  version: v1
  description: >-
    Stable service-to-service API for acquiring durable target leases and
    administering Backup Gateway. Backup traffic itself never passes through
    this API.
servers:
  - url: /
...
paths:
...
  /api/v1/admin/diagnostics/targets/{targetId}:
    parameters:
      - $ref: '#/components/parameters/TargetId'
    get:
      tags: [Diagnostics]
      operationId: getTargetDiagnostics
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Bounded operational target diagnostics.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TargetDiagnosticsResponse'
        '401': { description: Authentication required. }
        '403': { description: Administrator role required. }
        '404': { description: Configured target not found. }
  /api/v1/admin/diagnostics/audit:
    get:
      tags: [Diagnostics]
      operationId: getRecentAudit
      security:
        - bearerAuth: []
      parameters:
        - name: targetId
          in: query
          required: false
          schema:
            type: string
            maxLength: 128
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
      responses:
        '200':
          description: Recent append-only audit records, newest first.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AuditEventResponse'
        '400': { description: Invalid query bounds. }
        '401': { description: Authentication required. }
        '403': { description: Administrator role required. }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    TargetId:
      name: targetId
      in: path
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 128
        pattern: '^[a-z0-9](?:[a-z0-9._-]*[a-z0-9])?$'
...
  schemas:
...
    TargetLifecycleState:
      type: string
      enum: [Unknown, Offline, Starting, Online, Stopping, Faulted]
...
    HeldLeaseDiagnostic:
      type: object
      additionalProperties: false
      required: [leaseId, clientId, lastHeartbeatAtUtc, isStale]
      properties:
        leaseId: { type: string, format: uuid }
        clientId: { type: string, format: uuid }
        lastHeartbeatAtUtc: { type: string, format: date-time }
        isStale: { type: boolean }
    TargetDiagnosticsResponse:
      type: object
      additionalProperties: false
      required: [targetId, state, heldLeaseCount, staleLeaseCount, leasesTruncated, leases]
      properties:
        targetId: { type: string }
        state: { $ref: '#/components/schemas/TargetLifecycleState' }
        observedAtUtc: { type: [string, 'null'], format: date-time }
        heldLeaseCount: { type: integer, minimum: 0 }
        staleLeaseCount: { type: integer, minimum: 0 }
        leasesTruncated: { type: boolean }
        leases:
          type: array
          maxItems: 100
          items: { $ref: '#/components/schemas/HeldLeaseDiagnostic' }
    AuditEventResponse:
      type: object
      additionalProperties: false
      required: [id, occurredAtUtc, correlationId, eventType, outcome]
      properties:
        id: { type: string, format: uuid }
        occurredAtUtc: { type: string, format: date-time }
        correlationId: { type: string, format: uuid }
        actorClientId: { type: [string, 'null'], format: uuid }
        subjectClientId: { type: [string, 'null'], format: uuid }
        targetId: { type: [string, 'null'] }
        leaseId: { type: [string, 'null'], format: uuid }
        eventType: { type: string, maxLength: 64 }
        outcome: { type: string, maxLength: 32 }
        details: { type: [string, 'null'], maxLength: 1024 }